refactor: Main 클래스 분리
