+++ src/main/java/carRacing/CarRacingMain.java
@@ -1,0 +1,17 @@
+package carRacing;
+
+
+public class CarRacingMain {
+    public static void main(String[] args) {
+        InputView inputView = new InputView();
+
+        inputView.printNumberOfCars();
+        int numberOfCars = inputView.inputNumber();
+
+        inputView.printNumberOfRacing();
+        int numberOfRacing = inputView.inputNumber();
+
+        CarRacing racing = new CarRacing();
+        racing.start(numberOfCars, numberOfRacing);
+    }
+}
